+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## WIP
+
+### 🚀 Performance
+
+1. 添加bridge invoke性能检测逻辑(开发模式下默认开启).
+
 ## 2.6.5(2025-01-16)
 
 ### 🚀 Performance